Add trimming of file paths and String entries
Previously the files were stored in the JAR under the same path as they
were read from disk. If you specified the files to put in the JAR with
absolute paths, they would end up having the complete path like
/home/paddington/…/data.edn inside the JAR. That's not what we want of
course. Add the possibility to add files from absolute paths, but leave
off a prefix.

Also add the possibility of directly writing Strings to the JAR without
first writing to disk. I thought this would be necessary for the
pom.xml, but it turns out that I need a pom.xml separately on disk
anyway. So I'll change all that a bit.